大体完成 还有统计 筛选等小细节

+++ DataManageBackend/src/main/java/com/datamanagebackend/service/impl/section/SectionCrudServiceImpl.java
@@ -1,15 +1,21 @@
 package com.datamanagebackend.service.impl.section;
 
 import com.datamanagebackend.mapper.SectionMapper;
+import com.datamanagebackend.mapper.StudentSectionMapper;
+import com.datamanagebackend.pojo.StudentSection;
 import com.datamanagebackend.service.section.SectionCrudService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @Service
 public class SectionCrudServiceImpl implements SectionCrudService {
+    @Autowired
+    private StudentSectionMapper studentSectionMapper;
+
     @Autowired
     private SectionMapper sectionMapper;
 
@@ -44,6 +50,13 @@ public class SectionCrudServiceImpl implements SectionCrudService {
             return res;
         }
 
+        String _capacity = data.get("capacity");
+        if (_capacity == null || _capacity.trim().length() == 0) {
+            res.put("error_message", "课程容量不能为空！");
+            return res;
+        }
+        Integer capacity = Integer.parseInt(_capacity);
+
         String[] days = {"周一", "周二", "周三", "周四", "周五", "周六", "周日"};
         String[] hours = {"8:00-8:50", "9:00-9:50", "10:00-10:50", "11:00-11:50", "14:00-14:50", "15:00-15:50", "16:00-16:50", "17:00-17:50"};
         String[] options = section_time.split(" ");
@@ -51,6 +64,9 @@ public class SectionCrudServiceImpl implements SectionCrudService {
             res.put("error_message", "上课时间不能为空！");
             return res;
         }
+
+
+
         StringBuilder sectionTime = new StringBuilder();
         for (String option : options) {
             int x = Integer.parseInt(option);
@@ -59,7 +75,7 @@ public class SectionCrudServiceImpl implements SectionCrudService {
             sectionTime.append(day).append("/").append(hour).append("/");
         }
 
-        sectionMapper.insertSection(null, course_id, teacher_id, classroom_id, sectionTime.toString());
+        sectionMapper.insertSection(null, course_id, teacher_id, classroom_id, sectionTime.toString(), capacity);
 
         res.put("error_message", "success");
         return res;
@@ -70,9 +86,15 @@ public class SectionCrudServiceImpl implements SectionCrudService {
         Map<String, String> res = new HashMap<>();
 
         Integer section_id = Integer.parseInt(data.get("section_id"));
+
+        List<StudentSection> studentSectionList = studentSectionMapper.selectListBySectionId(section_id);
+        if (studentSectionList.size() != 0) {
+            res.put("error_message", "仍有选该课的学生！");
+            return res;
+        }
+
         sectionMapper.deleteSectionBySectionId(section_id);
         res.put("error_message", "success");
-
         return res;
     }
